refactor: rename directory name to kebab-case
